# spec/v2/swagger.yaml
...
swagger: '2.0'
info: 
  title: OpenBMP REST API
  description: Powering Dashboards with the OpenBMP REST API
  version: "1.0.0"
# the domain of the service
host: api.openbmp.org
...
schemes:
  - https
# will be prefixed to all paths
basePath: /v1
produces:
  - application/json
paths:
...
  /withdrawns/top:
    get:
      summary: Get the list of top prefixes by number of prefixes withdrawn in past 'hours'. Limit defaults to 25 and hours defaults to 2.
      description: This Withdrawns Endpoint returns the top prefixes by number of prefixes withdrawn in past 'hours'. Limit defaults to 25 and hours defaults to 2.
      parameters:
        - $ref: '#/parameters/searchPeer'
        - $ref: '#/parameters/searchPrefix'
        - $ref: '#/parameters/groupBy'
        - $ref: '#/parameters/joinWhoIsPrefix'
        - $ref: '#/parameters/limit'
        - $ref: '#/parameters/startTimestamp'
        - $ref: '#/parameters/endTimestamp'
      tags:
        - Withdrawns Over Time
      responses:
        200:
          description: An array of top prefixes by number of prefixes withdrawn in past 'hours'. Limit defaults to 25 and hours defaults to 2.
          schema:
            type: array
            items:
              $ref: '#/definitions/TopPrefixes'
        400:
          description: Bad Request - Missing Required Path Parameters
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized - No Authentication
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Forbidden - Authentication Correct, But You do not have persmission to perform this operation on this resource
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Resource Not Found
          schema:
            $ref: '#/definitions/Error'
        405:
          description: Method not allowed/invalid on this resource
          schema:
            $ref: '#/definitions/Error'
        500:
          description: OpenBMP Server Error
          schema:
            $ref: '#/definitions/Error'
        503:
          description: OpenBMP REST API Service temporarily unavailable due to maintenance
          schema:
            $ref: '#/definitions/Error'
        504:
          description: Gateway Timeout
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /withdrawns/peer/{peerHashId}/top:
    get:
      summary: For given peer, return the top prefixes by number of prefixes withdrawn in past 'hours'. Limit defaults to 25 and hours defaults to 2.
      description: |
        This Withdrawns Endpoint for a given peer, return the top prefixes by number of prefixes withdrawn in past 'hours'. Limit defaults to 25 and hours defaults to 2.
      parameters:
        - $ref: '#/parameters/peerHashId'
        - $ref: '#/parameters/limit'
        - $ref: '#/parameters/hours'
        - $ref: '#/parameters/timestamp'
      tags:
        - Withdrawns Over Time
      responses:
        200:
          description: For given peer, return the top prefixes by number of prefixes withdrawn in past 'hours'. Limit defaults to 25 and hours defaults to 2.
          schema:
            type: array
            items:
              $ref: '#/definitions/TopPrefixesByPeer'
        400:
          description: Bad Request - Missing Required Path Parameters
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized - No Authentication
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Forbidden - Authentication Correct, But You do not have persmission to perform this operation on this resource
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Resource Not Found
          schema:
            $ref: '#/definitions/Error'
        405:
          description: Method not allowed/invalid on this resource
          schema:
            $ref: '#/definitions/Error'
        500:
          description: OpenBMP Server Error
          schema:
            $ref: '#/definitions/Error'
        503:
          description: OpenBMP REST API Service temporarily unavailable due to maintenance
          schema:
            $ref: '#/definitions/Error'
        504:
          description: Gateway Timeout
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /withdrawns/top/interval/{minutes}:
    get:
      summary: Return the count of updates per interval for up to max limit size. Limit defaults to 25.
      description: |
        This Withdrawns Endpoint returns the count of updates per interval for up to max limit size. Limit defaults to 25.
      parameters:
        - $ref: '#/parameters/minutes'
        - $ref: '#/parameters/limit'
        - $ref: '#/parameters/timestamp'
      tags:
        - Withdrawns Over Time
      responses:
        200:
          description: Return the count of updates per interval for up to max limit size. Limit defaults to 25.
          schema:
            type: array
            items:
              $ref: '#/definitions/IntervalCount'
        400:
          description: Bad Request - Missing Required Path Parameters
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized - No Authentication
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Forbidden - Authentication Correct, But You do not have persmission to perform this operation on this resource
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Resource Not Found
          schema:
            $ref: '#/definitions/Error'
        405:
          description: Method not allowed/invalid on this resource
          schema:
            $ref: '#/definitions/Error'
        500:
          description: OpenBMP Server Error
          schema:
            $ref: '#/definitions/Error'
        503:
          description: OpenBMP REST API Service temporarily unavailable due to maintenance
          schema:
            $ref: '#/definitions/Error'
        504:
          description: Gateway Timeout
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /withdrawns/peer/{peerHashId}/top/interval/{minutes}:
    get:
      summary: For given peer, return the count of updates per interval for up to max limit size. Limit defaults to 25.
      description: |
        This Withdrawns Endpoint for a given peer, return the count of updates per interval for up to max limit size. Limit defaults to 25.
      parameters:
        - $ref: '#/parameters/peerHashId'
        - $ref: '#/parameters/minutes'
        - $ref: '#/parameters/limit'
        - $ref: '#/parameters/timestamp'
      tags:
        - Withdrawns Over Time
      responses:
        200:
          description: For given peer, return the count of updates per interval for up to max limit size. Limit defaults to 25.
          schema:
            type: array
            items:
              $ref: '#/definitions/IntervalCount'
        400:
          description: Bad Request - Missing Required Path Parameters
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized - No Authentication
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Forbidden - Authentication Correct, But You do not have persmission to perform this operation on this resource
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Resource Not Found
          schema:
            $ref: '#/definitions/Error'
        405:
          description: Method not allowed/invalid on this resource
          schema:
            $ref: '#/definitions/Error'
        500:
          description: OpenBMP Server Error
          schema:
            $ref: '#/definitions/Error'
        503:
          description: OpenBMP REST API Service temporarily unavailable due to maintenance
          schema:
            $ref: '#/definitions/Error'
        504:
          description: Gateway Timeout
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  GenericCount:
    title: Generic Count
    type: object
    properties:
      PrefixCount:
        type: integer
        format: int64
        description: Generic Count that could be used as Prefix Count or Prefix Learned etc.
...
  TopPrefixesByPeer:
    title: Top Prefixes By Peer
    type: object
    properties:
      Prefix:
        type: string
        description: Prefix 
      PrefixLen:
        type: string
        format: byte
        description: Prefix Length
      PeerName:
        type: string
        description: Peer Name
      PeerAddr:
        type: string
        description: Peer Address
      Count:
        type: object
        description : Count
        items:
          $ref: '#/definitions/GenericCount'
      peer_hash_id:
        type: string
        description: Peer Hash Id
      RouterName:
        type: string
        description: Router Name
  TopPrefixes:
    title: Top Prefixes
    type: object
    properties:
      top_prefixes_by_peer:
        type: object
        description: Top Prefixes by Peer Object
        items:
          $ref: '#/definitions/TopPrefixesByPeer'
      RouterAddr:
        type: string
        description: Router Address
      CollectorName:
        type: string
        description: Collector Name
      CollectorAddr:
        type: string
        description: Collector Address
      CollectorAdminID:
        type: string
        description: Collector Admin Id
      PrefixDescr:
        type: object
        description : Prefix Description Blob
        items:
          $ref: '#/definitions/BlobObject'
      OriginAS:
        type: integer
        format: int32
        description: Origin AS
  IntervalCount:
    title: Interval and Count
    type: object
    properties:
      IntervalTime:
        type: string
        description: Interval Time
      Count:
        type: object
        description : Count
        items:
          $ref: '#/definitions/GenericCount'
...
  BlobObject:
    title: Any BLOB Object
    type: object
    properties:
      prefix_descr:
        type: object
        description: Any BLOB Object
...
  Error:
    type: object
    properties:
      code:
        type: string
        description: Application specific error codes
        #
        # When adding a new error, add a new enum below as well as make corresponding entry in ErrorDetails
        #
        enum:
          - DATASOURCE_CONNECT_FAILURE
          - INVALID_PREFIX_BITS
          - VALIDATION_FAILED_TYPE1
          - VALIDATION_FAILED_TYPE2
          - REVERSE_PROXY_ERROR
          - LOAD_BALANCER_ERROR
          - NOT_READY_YET
          - USAGE_LIMIT_REACHED
          - FETCH_LIMIT_EXCEEDED
          - INVALID_COMMUNITY_VALUE
          - INVALID_P1
          - INVALID_HOSTNAME
          - INVALID_ASN
          - MISSING_REQD_PARAMETER_ASN
          - INVALID_PEERHASHID
          - MISSING_REQD_PARAMETER_PEERHASHID
          - INVALID_IPADDRESS
          - MISSING_REQD_PARAMETER_IP
          - MISSING_REQD_PARAMETER_LIMIT
          - MISSING_REQD_PARAMMETER_PAGE
          - INCORRECT_WHERE_CLAUSE
          - INCORRECT_SORT_VALUE
          - INCORRECT_SORT_DIRECTION_VALUE
          - INCORRECT_ORDER_BY_CLAUSE
          - INCORRECT_GROUP_BY_CLAUSE
          - INVALID_IP_START
          - INVALID_IP_END
          - COUNTRY_INVALID
          - STATE_PROV_INVALID
          - CITY_INVALID
          - INCORRECT_LAT_VALUE
          - INCORRECT_LONG_VALUE
          - INVALID_TIMEZONE
          - INVALID_TIMEZONE_OFFSET
          - INVALID_ISP
          - UNAVAILABLE_COLUMN
          - VALUE_SIZE_EXCEED_FOR_COLUMN
          - MISSING_REQD_PARAMETER_IP_START
          - MISSING_REQD_PARAMETER_COLUMN
          - MISSING_REQD_PARAMETER_VALUE
          - MISSING_REQD_PARAMETER_COUNTRY
          - MISSING_REQD_PARAMETER_CITY
          - INCORRECT_BOOLEAN_VALUE
          - MISSING_REQD_PARAMETER_MTID
          - INVALID_MTID
          - MISSING_REQD_PARAMETER_ROUTERID
          - INVALID_ROUTERID
          - MISSING_REQD_PARAMETER_LOCALIP
          - MISSING_REQD_PARAMETER_PEERIP
          - MISSING_REQD_PARAMETER_ROUTERIP
          - MISSING_REQD_PARAMETER_ROUTERHASHID
          - INVALID_ROUTERHASHID
          - MISSING_REQD_PARAMETER_PREFIX
          - MISSING_REQD_PARAMETER_LENGTH
          - INVALID_LENGTH
          - INVALID_HOURS
          - INVALID_TIMESTAMP
          - INVALID_PAGE
          - INVALID_SEARCHPEER
          - INVALID_SEARCHPREFIX
          - MISSING_REQD_PARAMETER_MINUTES
          - INVALID_MINUTES
          - MISSING_REQD_PARAMRTER_SECONDS
          - INVALID_SECONDS
          - MISSING_REQD_PARAMETER_PREFIX_LEN
          - INVALID_PREFIX_LEN
      message:
        type: string
      description:
        type: string
...
parameters:
  limit:
    name: limit
    in: query
    description: The Limit Value in SQL query
    type: integer
    format: int32
    required: false
...
  groupBy:
    name: groupBy
    in: query
    description: The Group By clause in SQL query
    type: string
    required: false
...
  hours:
    name: hours
    in: query
    description: The value in hours
    type: integer
    format: int32
    required: false
  timestamp:
    name: timestamp
    in: query
    description: The timestamp string
    type: string
    required: false
  startTimestamp:
    name: startTimestamp
    in: query
    description: The start timestamp string
    type: string
    required: false
  endTimestamp:
    name: endTimestamp
    in: query
    description: The end timestamp string
    type: string
    required: false
  searchPeer:
    name: searchPeer
    in: query
    description: The searchPeer value
    type: string
    required: false
  searchPrefix:
    name: searchPrefix
    in: query
    description: The searchPrefix value
    type: string
    required: false
  joinWhoIsPrefix:
    name: joinWhoIsPrefix
    in: query
    description: Boolean Flag
    type: boolean
    required: false
...
  peerHashId:
    name: peerHashId
    in: path
    description: The peerHashId value
    type: string
    required: true
...
  minutes:
    name: minutes
    in: path
    description: The value in minutes
    type: integer
    format: int32
    required: true
...